Fix C++ warnings

+++ examples/DoorSensor/DoorSensor.ino
@@ -3,7 +3,7 @@
 const int PIN_DOOR = 16;
 
 Bounce debouncer = Bounce(); // Bounce is built into Homie, so you can use it without including it first
-unsigned long lastDoorValue = -1;
+int lastDoorValue = -1;
 
 HomieNode doorNode("door", "door");
